+++ spec/spec_helper.rb
@@ -1,5 +1,6 @@
 # frozen_string_literal: true
 
+require 'byebug'
 require 'rubocop'
 require 'rubocop/rspec/support'
 require 'rubocop-ordered_methods'