+++ CHANGELOG.md
@@ -6,6 +6,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Autocorrection support for Sorbet signatures
+
 ## [0.8] - 2021-02-01
 
 ### Fixed